frontend/requirements.txt
streamlit>=1.32.0
httpx>=0.25.1
python-dotenv>=1.0.0 